+++ onsiteprint-plugin/CHANGELOG.md
@@ -45,6 +45,7 @@ This document includes the change log notes for the OnsitePrint Plugin - [GitHub
 - Added `opModalToggleListener()` and `opCreateParticipantListener()`.
 - Added the `Column Input Listener` Function.
 - Added new `Modal Function` to the `Modal Toggle Listener`.
+- Added new `Modal Toggle Listener` to the `CSV Download Button`.
     
     [EB | setup-header.js](blocks/event/block-script-parts/setup-header.js)